ffff

+++ problems/problem_1.cpp
@@ -1,4 +1,4 @@
-void problemSolution1(float consumed_water) {
+void problemSolution1(double consumed_water) {
     double cost;
         if (consumed_water > 60) {
             cost = (13 + 30*0.4 + 20*0.12 + 10*1.4 + (consumed_water-60)*1.5);
